requirement.txt
flask
flask-cors
authlib
python-dotenv

pandas
numpy

mysql-connector-python

werkzeug
requests
